Explicitly advise using UTC rather than local time
This method is the simplest and least error-prone
for experiments that span the start or end of daylight savings time.
I considered using 'date -v -1H'
to simulate CST (in file names) after the transition to CDT,
but that method is much harder to explain
and requires changing cron tables in the middle of an experiment.