One should calibrate chamber clock every time
Dmitri correctly pointed out to me
that adjusting the chamber clock
should be part of the standard procedure.
I had imagined calibrating it four times a year,
but that is more error-prone
(more complicated and harder to remember).

Hopefully this will encourage people
to experiment with light schedules,
given that they need to get growth facility staff
to alter the schedule anyway.