ReadMe: Abbreviate as 'RasPi' rather than 'raspi'
This is referring to the Raspberry Pi board.
The abbreviation 'RPi' (or just 'pi') is more common,
but looks worse to me when pluralized
('RPis', 'pis').

I still find two other phrases useful:
and 'RaspiCam' (for referring to the camera board)
'Pi/Camera rig' (for RasPi board, camera board, and case).